opt/web/panel/requirements.txt
# DebVisor Web Panel - Python Dependencies

# Web Framework
Flask>=3.0.3
Werkzeug>=3.0.3

# Authentication & Security
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Limiter>=3.7.0
cryptography>=42.0.8
PyJWT>=2.9.0
pyotp>=2.9.0
webauthn>=2.0.0

# Database
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.30

# Form handling
WTForms>=3.1.2

# gRPC for RPC service communication
grpcio>=1.64.0
grpcio-tools>=1.64.0
protobuf>=5.27.0

# Caching & sessions
redis>=5.0.4
python-dotenv>=1.0.1

# Production WSGI server
gunicorn>=22.0.0

# Utilities
bleach>=6.1.0
requests>=2.32.3

# Development & Testing
pytest>=8.2.0
pytest-flask>=1.3.0
black>=24.4.2
